Refactor seat viewer components and consolidate cloud configuration
